requirements.txt
transformers
datasets
torch
mlflow
fastapi
uvicorn
gradio
pandera
pydantic
slowapi
prometheus-fastapi-instrumentator
python-json-logger
bandit
pip-audit
detect-secrets
scikit-learn
pandas
numpy
psutil
httpx
pytest-asyncio
